Move frontend test utilities to a separate directory
This provides consistency with the Hypothesis client's organization.
See https://github.com/hypothesis/client/pull/1739.

The `test-util` dir name is hyphenated for the same reason.
See https://hypothes-is.slack.com/archives/C1M8NH76X/p1580827144128300.
