+++ lms/static/scripts/test-util/mock-imported-components.js
